User experience workflow in OCP CoreOS Layering model
Since the last time, we have gained experience while working
on the prototype in the layering branch https://github.com/openshift/machine-config-operator/tree/layering
and discussing/demoing to different teams.

Goal is to have up-to-date information of the existing
enhancement and sketch how the user experience will look like
in the layering model.